Create Largest Number
Q: Given a list of non-negative integers nums, arrange them such that they form the largest number and return it.

Since the result may be very large, so you need to return a string instead of an integer.

You should not change the position of the digits of any interger. If the given array nums is [20,1]. The numbers that can be formed are 201 and 120 and not 210.

Input Format:
single line input of space separated integers of nums.

Constraints:
1<=nums.length<=100
0<=nums[i]<=10 pow 9

Output Format: Single line string type output- of the largest formed number with the given inputs

After Compilation:-

Sample Input 0: 10 2
Sample Output 0: 210
Explanation 0: The numbers which can be formed by the given inputs[10,2] are 102 and 210. In this two numbers 210 is the largest one.

Sample Input 1: 3 30 34 5 9
Sample Output 1: 9534330
Explanation 1: 9534330 is the largest number formed using the inputs