Create Class Monitor
Q: A professor decides to deliver his seminar Infront of a class only if his required minimum number of students got present in the class before his deadline(Time).

You were provided with a couple of integer variables 'N' and 'M', Where N is the total number of students in the class and M is the number of students required for the lecturer to start the seminar. Along with a Deadline time(In String format), before which M number of students should be in the class. and a array of values indicating the arrival of each student is provided.

after the given inputs check the condition if M number of students got present in the class return a message as "Seminar will continue as per schedule", else return a message as "Seminar got cancelled"

After Compilation:-
N: 5
M: 3
t: '9:30'
l = ['9:30', '9:28', '9:29', '9:35', '9:37']

output: Seminar will continue as per schedule