Create Diagonal Summer
Q: You were provided with an integer 'n' which determines the number of rows and columns in a array(rows==columns). Enter the values of a multidimensional array based on its index and execute the program. The output will be the sum of main and sub diagonal of an array.

After Execution:-
Input: n = 3 | l = [[5,10,11], [79,6,12], [9,21,45]]

output: 76

Explanation:-
Sum of main diagonal + sum of sub diagonal : 5+6+45+9+11: 76